Keep CLI tests runnable across runtimes
Split the inbox and relay command parsers out of the runtime-heavy
modules so source-level CLI tests do not import .tsx or node:sqlite
under Node and Bun. This also rewrites the Deno version regression
test to replace the global safely and raises the Bun timeout for
vocab-tools so mise test stays green again.
