+++ packages/cli/src/runner.ts
@@ -8,12 +8,12 @@ import process from "node:process";
 import { parse as parseToml } from "smol-toml";
 import { configContext, tryLoadToml } from "./config.ts";
 import { generateVocabCommand } from "./generate-vocab/mod.ts";
-import { inboxCommand } from "./inbox.tsx";
+import { inboxCommand } from "./inbox/command.ts";
 import { initCommand } from "./init/mod.ts";
 import { lookupCommand } from "./lookup.ts";
 import { nodeInfoCommand } from "./nodeinfo.ts";
 import { globalOptions } from "./options.ts";
-import { relayCommand } from "./relay.ts";
+import { relayCommand } from "./relay/command.ts";
 import { tunnelCommand } from "./tunnel.ts";
 import { webFingerCommand } from "./webfinger/mod.ts";
 import metadata from "../deno.json" with { type: "json" };
